requirements.txt
atlassian-python-api
markdownify
jira
jira2markdown
networkx
python-magic
spacy
openai